+++ CHANGELOG.md
@@ -15,6 +15,7 @@ changed functionality, fixed bugs).
 - Worker can be told to shut down by the Manager. The environment (for example systemd
   on Linux) is responsible for restarting Flamenco Worker after such a shutdown.
 - Added `--version` CLI option to show the version of Flamenco Worker and quit.
+- Added `--single` or `-1` CLI option to shut down the Worker after executing a single task.
 
 
 ## Version 2.0.8 (released 2017-09-07)